+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+### Added
+- Longitudinal carriage model ceftriaxone vs all other abx
+
+### Changed
+- Correct typos throughout
+
 ## [1.1.4] - 2021-12-13
 ### Added
 - Table of Kleb isolates showing ESBL status and which study they came from